Move grad-free tests into testLocal()
